Add files via upload
This program is a Python script that evaluates the strength of passwords. It asks the user to enter one character at a time to form a password. With each added character, the program checks if the password is secure. A password is considered secure if it has at least 8 characters, contains at least one digit, one uppercase letter, one lowercase letter, and one special character. If the password is secure, the program displays a message indicating that the passwors is secure. Otherwise, it displays a message indicating that the password is not secure and encourages the user to try again.